Add files via upload
EDA stands for Exploratory Data Analysis. It is an approach to analyzing and visualizing data sets to summarize their key characteristics, often with the help of statistical graphics and other data visualization methods. The primary goal of EDA is to uncover patterns, trends, relationships, and anomalies within the data, helping to generate hypotheses and insights.
EDA is a crucial step in the data analysis process, providing a foundation for subsequent modelling and decision-making. It is widely used in various fields, including statistics, data science, and machine learning, to gain a deeper understanding of datasets before applying more advanced techniques.